+++ CHANGELOG.md
@@ -23,6 +23,7 @@
  - Dynmap will now show spell casts (enabled by default)
  - Allow wand randomization for levels past 40 - e.g. "80 levels" will add level-40 enchants twice.
  - Added "Column" (superiterate) spell.
+ - Added "Neutron Bomb" spell, make wither apply wither effects (MIGRATION NOTE)
  
 ## 2.8.9.1